Extract isEnumValue tests into core tests
